+++ gpu/cuda_common.cpp
@@ -10,7 +10,7 @@ SmallMtxArrayStatic::~SmallMtxArrayStatic(){
 }
 
 SmallMtxArrayStatic::SmallMtxArrayStatic(const std::vector<gpumat::GpuMat> &gmat, int beg, int last){
-	if(maxcount < last - beg || last > gmat.size() || beg > gmat.size())
+	if(maxcount < last - beg || last > (int)gmat.size() || beg > (int)gmat.size())
 		throw new std::invalid_argument("not enough size of array for store matrices");
 
 	count = last - beg;
